Refactor: RestManagementController
refactor la clase RestManagementController into RestManagementDestinationController and RestManagerItineraryController
