+++ TripScanner/src/main/java/com/tripscanner/TripScanner/controller/restController/RestManagementDestinationController.java
@@ -1,0 +1,84 @@
+package com.tripscanner.TripScanner.controller.restController;
+
+
+import com.tripscanner.TripScanner.model.Destination;
+import com.tripscanner.TripScanner.model.Place;
+import com.tripscanner.TripScanner.service.DestinationService;
+import com.tripscanner.TripScanner.service.ItineraryService;
+import org.hibernate.engine.jdbc.BlobProxy;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.sql.SQLException;
+import java.util.List;
+import java.util.Optional;
+
+import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentRequest;
+
+@RestController
+@RequestMapping("/api/management/")
+public class RestManagementDestinationController {
+    @Autowired
+    private ItineraryService itineraryService;
+
+    @Autowired
+    private DestinationService destinationService;
+
+
+    @GetMapping("/destination/")
+    public ResponseEntity<Destination> getDestination() {
+        List<Destination> destinations = destinationService.findAll();
+        if (!destinations.isEmpty()) {
+            return ResponseEntity.ok(destinations.get(1));
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
+
+    @PostMapping("/destination/")
+    public ResponseEntity<Destination> createDestination(@RequestBody Destination destination) {
+        destinationService.save(destination);
+        URI location = fromCurrentRequest().path("/{id}").buildAndExpand(destination.getId()).toUri();
+        return ResponseEntity.created(location).body(destination);
+    }
+
+    @PutMapping("/destination/{id}")
+    public ResponseEntity<Destination> editDestination(@PathVariable long id, @RequestBody Destination newDestination) throws SQLException {
+        Optional<Destination> destination = destinationService.findById(id);
+        if (destination.isPresent()) {
+            if (newDestination.getImageFile() != null) {
+                newDestination.setImageFile(BlobProxy.generateProxy(destination.get().getImageFile().getBinaryStream(), destination.get().getImageFile().length()));
+            }
+            newDestination.setId(id);
+            destinationService.save(newDestination);
+            return ResponseEntity.ok(destination.get());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
+
+    @DeleteMapping("/destination/{id}")
+    public ResponseEntity<Destination> deleteDestination(@PathVariable long id) {
+        Optional<Destination> destination = destinationService.findById(id);
+        if (destination.isPresent()) {
+            for (int i = 0; i < destination.get().getPlaces().size(); i++) {
+                Place place = destination.get().getPlaces().get(i);
+                for (int j = 0; j < place.getItineraries().size(); j++) {
+                    place.getItineraries().get(j).getPlaces().remove(place);
+                    itineraryService.findById(place.getItineraries().get(j).getId()).get().getPlaces().remove(place);
+                    if (itineraryService.findById(place.getItineraries().get(j).getId()).get().getPlaces().isEmpty()) {
+                        itineraryService.delete(place.getItineraries().get(j).getId());
+                    }
+                }
+            }
+            destinationService.delete(id);
+            return new ResponseEntity<>(null, HttpStatus.OK);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
+
+}
