+++ TripScanner/src/main/java/com/tripscanner/TripScanner/controller/restController/RestManagementItineraryController.java
@@ -1,11 +1,9 @@
 package com.tripscanner.TripScanner.controller.restController;
 
 
-import com.tripscanner.TripScanner.model.Destination;
 import com.tripscanner.TripScanner.model.Itinerary;
 import com.tripscanner.TripScanner.model.Place;
 import com.tripscanner.TripScanner.model.User;
-import com.tripscanner.TripScanner.service.DestinationService;
 import com.tripscanner.TripScanner.service.ItineraryService;
 import com.tripscanner.TripScanner.service.UserService;
 import org.hibernate.engine.jdbc.BlobProxy;
@@ -18,87 +16,20 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
-import java.net.URI;
 import java.security.Principal;
-import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentRequest;
-
 @RestController
 @RequestMapping("/api/management/")
-
-public class RestManagementController {
+public class RestManagementItineraryController {
     @Autowired
     private ItineraryService itineraryService;
 
     @Autowired
-    private DestinationService destinationService;
-
-    @Autowired
     private UserService userService;
 
-
-    //destination
-
-    @GetMapping("/destinations/")
-    public ResponseEntity<Destination> getDestination() {
-        List<Destination> destinations = destinationService.findAll();
-        if (!destinations.isEmpty()) {
-            return ResponseEntity.ok(destinations.get(1));
-        } else {
-            return ResponseEntity.notFound().build();
-        }
-    }
-
-    @PostMapping("/destinations/")
-    public ResponseEntity<Destination> createDestination(@RequestBody Destination destination) {
-        destinationService.save(destination);
-        URI location = fromCurrentRequest().path("/{id}").buildAndExpand(destination.getId()).toUri();
-        return ResponseEntity.created(location).body(destination);
-    }
-
-    @PutMapping("/destinations/{id}")
-    public ResponseEntity<Destination> editDestination(@PathVariable long id, @RequestBody Destination newDestination) throws SQLException {
-        Optional<Destination> destination = destinationService.findById(id);
-        if (destination.isPresent()) {
-            if (newDestination.getImageFile() != null) {
-                newDestination.setImageFile(BlobProxy.generateProxy(destination.get().getImageFile().getBinaryStream(), destination.get().getImageFile().length()));
-            }
-            newDestination.setId(id);
-            destinationService.save(newDestination);
-            return ResponseEntity.ok(destination.get());
-        } else {
-            return ResponseEntity.notFound().build();
-        }
-    }
-
-    @DeleteMapping("/destinations/{id}")
-    public ResponseEntity<Destination> deleteDestination(@PathVariable long id) {
-        Optional<Destination> destination = destinationService.findById(id);
-        if (destination.isPresent()) {
-            for (int i = 0; i < destination.get().getPlaces().size(); i++) {
-                Place place = destination.get().getPlaces().get(i);
-                for (int j = 0; j < place.getItineraries().size(); j++) {
-                    place.getItineraries().get(j).getPlaces().remove(place);
-                    itineraryService.findById(place.getItineraries().get(j).getId()).get().getPlaces().remove(place);
-                    if (itineraryService.findById(place.getItineraries().get(j).getId()).get().getPlaces().isEmpty()) {
-                        itineraryService.delete(place.getItineraries().get(j).getId());
-                    }
-                }
-            }
-            destinationService.delete(id);
-            return new ResponseEntity<>(null, HttpStatus.OK);
-        } else {
-            return ResponseEntity.notFound().build();
-        }
-    }
-
-
-    //itinerary
-    @GetMapping("/itineraries/")
+    @GetMapping("/itinerary/")
     public ResponseEntity<Itinerary> getItinerary() {
         List<Itinerary> itinerary = itineraryService.findAll();
         if (!itinerary.isEmpty()) {
@@ -108,7 +39,7 @@ public class RestManagementController {
         }
     }
 
-    @PostMapping("/itineraries/")
+    @PostMapping("/itinerary/")
     public ResponseEntity<Long> createNewItinerary(@RequestBody Itinerary itinerary, HttpServletRequest request) throws IOException {
         Principal principalUser = request.getUserPrincipal();
         if (principalUser == null) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -129,7 +60,7 @@ public class RestManagementController {
     }
 
 
-    @PutMapping("/itineraries/{id}")
+    @PutMapping("/itinerary/{id}")
     public ResponseEntity editItinerary(@PathVariable long id, @RequestBody Itinerary newItineraries, HttpServletRequest request) {
         Optional<Itinerary> itinerary = itineraryService.findById(id);
 
@@ -152,7 +83,7 @@ public class RestManagementController {
         }
     }
 
-    @DeleteMapping("/itineraries/{id}")
+    @DeleteMapping("/itinerary/{id}")
     public ResponseEntity deleteItinerary(@PathVariable long id) {
         Optional<Itinerary> itinerary = itineraryService.findById(id);
         if (itinerary.isPresent()) {
